STUDIO-10369: Fixing case when root element from JSON schema is a reference
