+++ src/main/java/org/mule/common/metadata/JSONSchemaMetadataModelFactory.java
@@ -8,6 +8,8 @@ import org.mule.common.metadata.parser.json.*;
 
 import java.io.IOException;
 import java.net.URL;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  *
@@ -20,9 +22,11 @@ public class JSONSchemaMetadataModelFactory
     public static final String TYPE = "type";
     public static final String PROPERTIES = "properties";
 
+    private final JSONMetadataModelFactoryHelper helper;
+
     public JSONSchemaMetadataModelFactory()
     {
-
+        this.helper = new JSONMetadataModelFactoryHelper();
     }
 
     private MetaDataModel buildModel(String jsonSchemaString, URL jsonSchemaURL)
@@ -37,7 +41,7 @@ public class JSONSchemaMetadataModelFactory
                 final JSONType itemsType = arrayType.getItemsType();
                 if (itemsType.isJSONObject())
                 {
-                    final DefaultStructuredMetadataModel model = new DefaultStructuredMetadataModel(DataType.JSON, new JSONSchemaMetaDataFieldFactory((JSONObjectType) itemsType));
+                    final DefaultStructuredMetadataModel model = new DefaultStructuredMetadataModel(DataType.JSON, new JSONSchemaMetaDataFieldFactory((JSONObjectType) itemsType, helper));
                     return new DefaultListMetaDataModel(model);
                 }
                 else if (itemsType.isJSONPrimitive())
@@ -53,9 +57,14 @@ public class JSONSchemaMetadataModelFactory
                 }
 
             }
+            else if (new SchemaEnv().evaluate(jsonSchemaObject).isJSONPointer())
+            {
+                final JSONPointerType pointer = (JSONPointerType) new SchemaEnv(null, jsonSchemaObject).evaluate(jsonSchemaObject);
+                return helper.buildJSONPointerMetaDataModel(pointer);
+            }
             else if ((jsonSchemaObject.has(TYPE) && jsonSchemaObject.get(TYPE).toString().toLowerCase().equals(OBJECT_ELEMENT_NAME)) || jsonSchemaObject.has(PROPERTIES))
             {
-                final JSONSchemaMetaDataFieldFactory fieldFactory = new JSONSchemaMetaDataFieldFactory(new JSONObjectType(new SchemaEnv(jsonSchemaObject, jsonSchemaURL), jsonSchemaObject));
+                final JSONSchemaMetaDataFieldFactory fieldFactory = new JSONSchemaMetaDataFieldFactory(new JSONObjectType(new SchemaEnv(jsonSchemaObject, jsonSchemaURL), jsonSchemaObject), helper);
                 return new DefaultStructuredMetadataModel(DataType.JSON, fieldFactory);
             }
             else if (new SchemaEnv().evaluate(jsonSchemaObject).isJSONPrimitive())
@@ -86,5 +95,55 @@ public class JSONSchemaMetadataModelFactory
     {
         String jsonSchemaString = IOUtils.toString(url.openStream());
         return buildModel(jsonSchemaString, url);
+    }
+
+    protected class JSONMetadataModelFactoryHelper {
+
+        private final Map<JSONObjectType, DefaultStructuredMetadataModel> visitedTypes = new HashMap<>();
+
+        public AbstractMetaDataModel buildJSONArrayMetaDataModel(JSONArrayType property) {
+            AbstractMetaDataModel model = null;
+            JSONType itemsType = property.getItemsType();
+            if (itemsType.isJSONPrimitive()) { // Case List<String>
+                DataType dataType = JSONTypeUtils.getDataType(itemsType);
+                model = dataType == DataType.UNKNOWN ? new DefaultUnknownMetaDataModel() : new DefaultSimpleMetaDataModel(dataType);
+            } else {
+                if(itemsType.isJSONPointer()){
+                    model = buildJSONPointerMetaDataModel((JSONPointerType) itemsType);
+                }else if (itemsType.isJSONObject()){
+                    model = buildJSONMetaDataModel((JSONObjectType) itemsType);
+                }
+            }
+            return model;
+        }
+
+        private AbstractMetaDataModel buildJSONPointerMetaDataModel(JSONPointerType pointer) {
+            JSONType resolvedType = pointer.resolve();
+            if (resolvedType.isJSONArray()) {
+                return buildJSONArrayMetaDataModel((JSONArrayType) resolvedType);
+            } else if (resolvedType.isJSONObject()) {
+                return buildJSONMetaDataModel((JSONObjectType) resolvedType);
+            } else if (resolvedType.isJSONPointer()) {
+                return buildJSONPointerMetaDataModel((JSONPointerType) resolvedType);
+            } else if (resolvedType.isJSONPrimitive()) {
+                DataType dataType = JSONTypeUtils.getDataType(resolvedType);
+                return dataType == DataType.UNKNOWN ? new DefaultUnknownMetaDataModel() : new DefaultSimpleMetaDataModel(dataType);
+            }
+            return null;
+        }
+
+        public DefaultStructuredMetadataModel buildJSONMetaDataModel(JSONObjectType type) {
+
+            DefaultStructuredMetadataModel model;
+            if (visitedTypes.containsKey(type)) {
+                model = visitedTypes.get(type);
+            } else {
+                model = new DefaultStructuredMetadataModel(DataType.JSON);
+                visitedTypes.put(type, model);
+                model.loadFieldsFrom(new JSONSchemaMetaDataFieldFactory(type, this));
+            }
+            return model;
+        }
+
     }
 }
